server-python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
tidalapi>=0.8.11
python-dotenv>=1.0.1
anyio>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-mock>=3.12.0
